mx-kz-ytml-2/kz-ytml-1/wz-hfmcp-1/WX/curs/cur1/pyproject.toml
[project]
name = "c1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio[mcp]>=5.32.0",
    "huggingface-hub>=0.32.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "smolagents>=1.17.0",
]
